fix: Check if the starting time is nil
It is possible that the ending time is non-nil but the starting time is nil. It
can happen for a block that has already started but not finished yet. In such a
case, if the block has a duration, the ending time will be calculated from the
started time and the duration, even without the starting time.